+++ src/main/java/kitchenpos/products/domain/ProductPrice.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
