+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.domain;
 
+
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -13,4 +14,3 @@ public interface ProductRepository {
 
     List<Product> findAllByIdIn(List<UUID> ids);
 }
-
