+++ src/main/java/kitchenpos/products/infra/DefaultPurgomalumClient.java
@@ -1,11 +1,12 @@
 package kitchenpos.products.infra;
 
+import java.net.URI;
+
+import kitchenpos.products.domain.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
 
 @Component
 public class DefaultPurgomalumClient implements PurgomalumClient {
