+++ src/main/java/kitchenpos/products/infra/JpaProductRepository.java
@@ -1,9 +1,9 @@
-package kitchenpos.products.tobe.infra;
+package kitchenpos.products.infra;
 
 import java.util.UUID;
 
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
