--- src/main/java/kitchenpos/products/tobe/application/ProductService.java
@@ -1,70 +1,0 @@
-package kitchenpos.products.tobe.application;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.UUID;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.tobe.domain.PurgomalumClient;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class ProductService {
-    private final ProductRepository productRepository;
-    private final MenuRepository menuRepository;
-    private final PurgomalumClient purgomalumClient;
-
-    public ProductService(
-        final ProductRepository productRepository,
-        final MenuRepository menuRepository,
-        final PurgomalumClient purgomalumClient
-    ) {
-        this.productRepository = productRepository;
-        this.menuRepository = menuRepository;
-        this.purgomalumClient = purgomalumClient;
-    }
-
-    @Transactional
-    public ProductResponse create(final ProductRequest productRequest) {
-        final Product product = productRepository.save(
-                Product.create(
-                        productRequest.getName(),
-                        productRequest.getPrice(),
-                        purgomalumClient)
-        );
-        return ProductResponse.of(product);
-    }
-
-    @Transactional
-    public ProductResponse changePrice(final UUID productId, final ProductRequest productRequest) {
-        final Product product = productRepository.findById(productId)
-            .orElseThrow(NoSuchElementException::new);
-        product.changePrice(productRequest.getPrice());
-        final List<Menu> menus = menuRepository.findAllByProductId(productId);
-        for (final Menu menu : menus) {
-            BigDecimal sum = BigDecimal.ZERO;
-            for (final MenuProduct menuProduct : menu.getMenuProducts()) {
-                sum = sum.add(
-                    menuProduct.getProduct()
-                        .getPrice()
-                        .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
-                );
-            }
-            if (menu.getPrice().compareTo(sum) > 0) {
-                menu.setDisplayed(false);
-            }
-        }
-        return ProductResponse.of(product);
-    }
-
-    @Transactional(readOnly = true)
-    public List<ProductResponse> findAll() {
-        List<Product> products = productRepository.findAll();
-        return products.stream().map(ProductResponse::of).toList();
-    }
-}
