--- src/main/java/kitchenpos/products/tobe/domain/PurgomalumClient.java
@@ -1,5 +1,0 @@
-package kitchenpos.products.tobe.domain;
-
-public interface PurgomalumClient {
-    boolean containsProfanity(String text);
-}
