+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -2,7 +2,8 @@ package kitchenpos.products.application;
 
 
 
-import kitchenpos.products.infra.PurgomalumClient;
+
+import kitchenpos.products.domain.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
