+++ src/test/java/kitchenpos/products/tobe/Fixtures.java
@@ -14,8 +14,8 @@ import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuGroup;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.products.tobe.application.FakePurgomalumClient;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.PurgomalumClient;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.PurgomalumClient;
 
 public class Fixtures {
     public static final UUID INVALID_ID = new UUID(0L, 0L);
