+++ src/test/java/kitchenpos/products/tobe/application/ProductServiceTest.java
@@ -10,8 +10,11 @@ import java.util.List;
 import java.util.UUID;
 import kitchenpos.menus.application.InMemoryMenuRepository;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.tobe.domain.PurgomalumClient;
+import kitchenpos.products.application.ProductRequest;
+import kitchenpos.products.application.ProductResponse;
+import kitchenpos.products.application.ProductService;
+import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
